:truck: rename api/channel.ts to api.ts and update imports
